add 403, 'Non hai il permesso di accedere, il tuo URL è stato modificato'

+++ app/Http/Controllers/Admin/DishController.php
@@ -10,28 +10,22 @@ use Illuminate\Support\Facades\Auth;
 
 class DishController extends Controller
 {
-    
     private $validations = [
         'name' => "required|string|max:255",
         'description' => "required|string|max:2500",
         'ingredients' => "required|string|max:255",
         'price' => "required|numeric|max:255",
-
     ];
+
     private $validations_messages = [
-        'required' => 'il campo :attribute è obbligatorio',
-        'min' => 'il campo :attribute deve avere minimo :min caratteri',
-        'max' => 'il campo :attribute non può superare i :max caratteri',
-        'url' => 'il campo deve essere un url valido',
-        'exists' => 'Valore non valido'
+        'required' => 'Il campo :attribute è obbligatorio',
+        'max' => 'Il campo :attribute non può superare i :max caratteri',
     ];
 
-    
     public function index()
     {
-        $user = Auth::user(); 
-        $restaurant = $user->restaurant; 
-    
+        $user = Auth::user();
+        $restaurant = $user->restaurant;
         $dishes = $restaurant->dishes;
         return view('admin.dishes.index', compact('dishes', 'restaurant'));
     }
@@ -47,13 +41,10 @@ class DishController extends Controller
 
         $data = $request->all();
 
-        // dd($data);
-        
-        // Recupera l'ID del ristorante associato all'utente autenticato
         $restaurantId = auth()->user()->restaurant->id;
-        
+
         $newDish = new Dish();
-        
+
         $visible = $request->has('visible') ? true : false;
 
         $newDish->name = $data['name'];
@@ -61,9 +52,8 @@ class DishController extends Controller
         $newDish->ingredients = $data['ingredients'];
         $newDish->visible = $visible;
         $newDish->price = $data['price'];
-        $newDish->photo       = $data['photo'];
+        $newDish->photo = $data['photo'];
 
-        // Assegna l'ID del ristorante al nuovo piatto
         $newDish->restaurant_id = $restaurantId;
 
         $newDish->save();
@@ -71,23 +61,33 @@ class DishController extends Controller
         return redirect()->route('admin.restaurants.index', ['dish' => $newDish]);
     }
 
-   
     public function show($id)
     {
         $dish = Dish::findOrFail($id);
-
+        
+        if (Auth::user()->id !== $dish->restaurant->user_id) {
+            return abort(403, 'Non hai il permesso di accedere, il tuo URL è stato modificato');
+        }
+        
         return view('admin.dishes.show', compact('dish'));
     }
 
-  
     public function edit(Dish $dish)
     {
+        if (Auth::user()->id !== $dish->restaurant->user_id) {
+            return abort(403, 'Non hai il permesso di accedere, il tuo URL è stato modificato');
+        }
+
         return view('admin.dishes.edit', compact('dish'));
     }
 
     public function update(Request $request, Dish $dish)
     {
         $request->validate($this->validations, $this->validations_messages);
+
+        if (Auth::user()->id !== $dish->restaurant->user_id) {
+            return abort(403, 'Non hai il permesso di accedere, il tuo URL è stato modificato');
+        }
 
         $data = $request->all();
 
@@ -98,18 +98,21 @@ class DishController extends Controller
         $dish->ingredients = $data['ingredients'];
         $dish->visible = $visible;
         $dish->price = $data['price'];
-
+        $dish->photo = $data['photo'];
 
         $dish->update();
 
         return redirect()->route('admin.restaurants.index', ['dish' => $dish->id]);
     }
 
-  
     public function destroy(Dish $dish)
     {
+        if (Auth::user()->id !== $dish->restaurant->user_id) {
+            return abort(403, 'Non hai il permesso di accedere, il tuo URL è stato modificato');
+        }
+
         $dish->delete();
-        
+
         return redirect()->route('admin.restaurants.index')->with('delete_success', $dish);
     }
 }
